# swagger/static/openapi.yaml
openapi: 3.0.0
info:
  version: 0.0.1
  title: Flask RestAPI
  description: >
    A Python web server that provides login/register rest api functionalities based on Flask freamwork.
servers:
  - url: http://localhost:5000
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: User login.
      operationId: login
      requestBody:
        description: Login required fields
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginPayload"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Response"
  /auth/register:
    post:
      tags:
        - Auth
      summary: New user registeration.
      operationId: register
      requestBody:
        description: Register required fields
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPayload"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Response"
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    LoginPayload:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - username
        - email
        - password
    RegisterPayload:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
      required:
        - username
        - email
    Response:
      type: object
      properties:
        message:
          type: string
        data:
          type: string
      required:
        - message
...